requirements.txt
django==4.2.24
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1